requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.3
certifi==2023.5.7
charset-normalizer==3.1.0
cytoolz==0.12.1
Django==4.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-tester==0.9.0b1
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
lru-dict==1.1.8
multidict==6.0.4
packaging==23.1
parsimonious==0.9.0
pluggy==1.0.0
protobuf==4.23.1
psycopg2-binary==2.9.6
pycryptodome==3.18.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
rlp==3.0.0
semantic-version==2.10.0
sqlparse==0.4.4
toolz==0.12.0
uritemplate==4.1.1
urllib3==2.0.2
web3==6.4.0
websockets==11.0.3
yarl==1.9.2
